requirements.txt
aiohttp==3.7.4.post0
asgiref==3.3.4
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.6
chardet==4.0.0
discord.py==1.7.1
Django==3.2
djangorestframework==3.12.4
idna==3.1
multidict==5.1.0
pycodestyle==2.7.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
yarl==1.6.3
